feat(appium): configuration for driver schemas
With this change, a driver can now specify a `schema` property in its config, which points to a `.json`, `.js` or `.cjs` schema.  Any schema which [Ajv](https://npm.im/ajv) can read works fine.

If the schema is present and valid, it will be registered with the internal `Ajv` schema cache, where we can use it to validate options (e.g., `driverArgs`).

Added some basic unit tests for `driver-config` since I modified it.

Regarding whether the unique schema identifier should be the `pkgName` or not: I'm not sure. `automationName` may be fine as well.  We will use this ID to retrieve the schema from the `Ajv` cache and validate some object against it.