fix(gulp-plugins): post-transpile tasks always run after transpile
They were only being run after `transpile` in a test context, but they should always be run after `transpile`.

`transpile` is now a series of tasks, the first being `transpileSources` (which was what `transpile` did before), and the second being task `post-transpile` which contains tasks to run in parallel, if any were defined in `opts.postTranspile`.